refactor(svelte): split plot-surface-inspection by host call path
Carve the 549-line pure inspection policy module into frame (queue/frame
routing), apply (setInspection/pin/mode), and teardown (blur/reconcile/emit/
dismiss) files. Update GGPlot and unit tests to import each path directly;
delete the monolithic module (no barrel). Behavior-preserving; tests green.
